Fixes [WIP]

+++ gaiku-3d/src/formats/gox.rs
@@ -30,7 +30,7 @@ impl FileFormat for GoxReader {
                             for data in layer.blocks.iter() {
                                 let block = block_data[data.block_index];
                                 let mut chunk = Chunk::new(
-                                    [data.x as f64, data.y as f64, data.z as f64],
+                                    [data.x as f32, data.y as f32, data.z as f32],
                                     16,
                                     16,
                                     16,
